diagonal movement and engine refactor
