+++ src/scripts/shared/engine.ts
@@ -2,7 +2,11 @@ import RAPIER from "@dimforge/rapier2d"
 import * as PIXI from "pixi.js"
 import useEngine from "../../composeables/use-engine"
 import { Entity } from "./entities/entity"
+import { RenderSystem } from "./systems/render"
+import { WorldMap } from "./util/create-world"
+import { LocalStorageManager } from "./util/local-storage-manager"
 import { SaveManager } from "./util/save-manager"
+import { World } from "./util/world"
 
 export class Engine {
 
@@ -12,6 +16,7 @@ export class Engine {
     paused: boolean = false
     running: boolean = false
     shifting: boolean = false
+    localStorageManager!: LocalStorageManager<any>
     saveManager: SaveManager = new SaveManager()
     textStyle: PIXI.TextStyle = new PIXI.TextStyle({
         fontFamily: 'PixiPressStart2P',
@@ -20,6 +25,9 @@ export class Engine {
     })
 
     player!: Entity
+    world!: World
+    worldIndex: number = 0
+    worlds!: Array<WorldMap>
 
     constructor(run?: boolean) {
         // set app
@@ -51,7 +59,9 @@ export class Engine {
     }
 
     addCanvas(elementRef: HTMLElement) {
-        // Override this method in each subclass
+        const render = this.world.getSystemByType('render') as RenderSystem
+        if (render) render.appendElement(elementRef)
+        else console.log("still loading.")
     }
 
     pause(): void {
@@ -67,6 +77,16 @@ export class Engine {
 
     resumeTicker(): void {
         this.app.ticker.start()
+    }
+
+    save(): void {
+        this.paused = true
+        this.localStorageManager.clearData()
+        this.saveManager.clearData()
+        this.worlds[this.worldIndex] = this.saveManager.createWorldMap(this.world)
+        this.saveManager.setData(this.worldIndex, this.worlds)
+        this.localStorageManager.saveData(this.saveManager.data)
+        this.pause()
     }
 
     startRun(): void {
